# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Geem CRM & E-Commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /brands:
    get:
      operationId: listBrands
      tags: [brands]
      summary: List brands
      responses:
        "200":
          description: Brands
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Brand"
    post:
      operationId: createBrand
      tags: [brands]
      summary: Create brand
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrandInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Brand"
  /brands/{id}:
    get:
      operationId: getBrand
      tags: [brands]
      summary: Get brand
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Brand
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Brand"
    patch:
      operationId: updateBrand
      tags: [brands]
      summary: Update brand
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrandUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Brand"
    delete:
      operationId: deleteBrand
      tags: [brands]
      summary: Delete brand
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- MODELS ---
  /models:
    get:
      operationId: listModels
      tags: [models]
      summary: List models
      parameters:
        - in: query
          name: brandId
          schema:
            type: integer
      responses:
        "200":
          description: Models
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeviceModel"
    post:
      operationId: createModel
      tags: [models]
      summary: Create model
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceModelInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceModel"
  /models/{id}:
    patch:
      operationId: updateModel
      tags: [models]
      summary: Update model
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceModelUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceModel"
    delete:
      operationId: deleteModel
      tags: [models]
      summary: Delete model
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- CATEGORIES ---
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List categories
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update category
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete category
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- VENDORS ---
  /vendors:
    get:
      operationId: listVendors
      tags: [vendors]
      summary: List vendors
      responses:
        "200":
          description: Vendors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vendor"
    post:
      operationId: createVendor
      tags: [vendors]
      summary: Create vendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/{id}:
    get:
      operationId: getVendor
      tags: [vendors]
      summary: Get vendor
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
    patch:
      operationId: updateVendor
      tags: [vendors]
      summary: Update vendor
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
    delete:
      operationId: deleteVendor
      tags: [vendors]
      summary: Delete vendor
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- COURIERS ---
  /couriers:
    get:
      operationId: listCouriers
      tags: [couriers]
      summary: List couriers
      responses:
        "200":
          description: Couriers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Courier"
    post:
      operationId: createCourier
      tags: [couriers]
      summary: Create courier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourierInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Courier"
  /couriers/{id}:
    patch:
      operationId: updateCourier
      tags: [couriers]
      summary: Update courier
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourierUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Courier"
    delete:
      operationId: deleteCourier
      tags: [couriers]
      summary: Delete courier
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- PAYMENT METHODS ---
  /payment-methods:
    get:
      operationId: listPaymentMethods
      tags: [paymentMethods]
      summary: List payment methods
      responses:
        "200":
          description: Payment methods
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentMethod"
    post:
      operationId: createPaymentMethod
      tags: [paymentMethods]
      summary: Create payment method
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentMethodInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentMethod"
  /payment-methods/{id}:
    patch:
      operationId: updatePaymentMethod
      tags: [paymentMethods]
      summary: Update payment method
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentMethodUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentMethod"
    delete:
      operationId: deletePaymentMethod
      tags: [paymentMethods]
      summary: Delete payment method
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- CUSTOMERS ---
...
components:
  schemas:
...
    Brand:
      type: object
      required: [id, name, active]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        deviceIdMandatory:
          type: boolean
        active:
          type: boolean
        modelsCount:
          type: integer
        createdAt:
          type: string
    BrandInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        deviceIdMandatory:
          type: boolean
        active:
          type: boolean
    BrandUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        deviceIdMandatory:
          type: boolean
        active:
          type: boolean

    # --- MODELS ---
    DeviceModel:
      type: object
      required: [id, brandId, name, active]
      properties:
        id:
          type: integer
        brandId:
          type: integer
        brandName:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        hasImei:
          type: boolean
        hasDeviceId:
          type: boolean
        hasIccid:
          type: boolean
        hasMsisdn:
          type: boolean
        deviceIdMandatory:
          type: boolean
        warrantyDays:
          type: integer
        active:
          type: boolean
        createdAt:
          type: string
    DeviceModelInput:
      type: object
      required: [brandId, name]
      properties:
        brandId:
          type: integer
        name:
          type: string
        description:
          type: string
        hasImei:
          type: boolean
        hasDeviceId:
          type: boolean
        hasIccid:
          type: boolean
        hasMsisdn:
          type: boolean
        deviceIdMandatory:
          type: boolean
        warrantyDays:
          type: integer
        active:
          type: boolean
    DeviceModelUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        warrantyDays:
          type: integer
        active:
          type: boolean

    # --- CATEGORIES ---
    Category:
      type: object
      required: [id, name, active]
      properties:
        id:
          type: integer
        name:
          type: string
        parentId:
          type: ["integer", "null"]
        parentName:
          type: ["string", "null"]
        active:
          type: boolean
        productsCount:
          type: integer
    CategoryInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        parentId:
          type: integer
        active:
          type: boolean
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
        parentId:
          type: integer
        active:
          type: boolean

    # --- VENDORS ---
    Vendor:
      type: object
      required: [id, name, active]
      properties:
        id:
          type: integer
        name:
          type: string
        contactPerson:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        taxNumber:
          type: ["string", "null"]
        paymentTerms:
          type: ["string", "null"]
        ledgerBalance:
          type: number
        active:
          type: boolean
        createdAt:
          type: string
    VendorInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        contactPerson:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        taxNumber:
          type: string
        paymentTerms:
          type: string
    VendorUpdate:
      type: object
      properties:
        name:
          type: string
        contactPerson:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        active:
          type: boolean

    # --- COURIERS ---
    Courier:
      type: object
      required: [id, name, active]
      properties:
        id:
          type: integer
        name:
          type: string
        apiProvider:
          type: ["string", "null"]
        active:
          type: boolean
        ledgerBalance:
          type: number
        createdAt:
          type: string
    CourierInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        apiProvider:
          type: string
        apiKey:
          type: string
        apiPassword:
          type: string
        active:
          type: boolean
    CourierUpdate:
      type: object
      properties:
        name:
          type: string
        active:
          type: boolean

    # --- PAYMENT METHODS ---
    PaymentMethod:
      type: object
      required: [id, name, type, active]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        active:
          type: boolean
        accountDetails:
          type: ["string", "null"]
    PaymentMethodInput:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        type:
          type: string
        active:
          type: boolean
        accountDetails:
          type: string
    PaymentMethodUpdate:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        active:
          type: boolean
        accountDetails:
          type: string

    # --- CUSTOMERS ---
...